Refactor import paths and clean up unused files
- Updated import paths for `refreshCacheInBackground` and `getRefreshStatus` to reflect their new locations in the `homepage` directory.
- Refactored import paths for `generateFreshHomepage` and related functions to align with the new structure.
- Deleted unused files related to background refresh, cache, and clustering services to streamline the codebase and improve maintainability.
