+++ src/lib/news/rss-feeds.json
